requirements.txt
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
python-dotenv
chromadb
tiktoken
pypdf
